Add matype to dvbs_muxlist

+++ src/neumodb/chdb/chdb_extra.h
@@ -235,7 +235,7 @@ namespace chdb {
 		return s;
 	}
 
-	void matype_str(ss::string_& s, uint8_t matype);
+	void matype_str(ss::string_& s, int16_t matype);
 	inline auto matype_str(uint8_t matype) {
 		ss::string<32> s;
 		matype_str(s, matype);
